[CR] Compute close time matches
Just truncating fails when execution spans a rounding boundary.  Meanwhile comparing test time
to test DB time is reasonably safe because the DB container is started as part of the test -- no
significant time skew is expected.  (But yes, a test that loads a database is a component rather
than a system test.)