Use regexp array directly in LastKeysInPrefixRegexp
It is not possible to roundtrip generating and unparsing disjunctions from the database: the
regexp parser optimizes and rewrites portions of the regexp.  So we need a copy of the input
array stored in the database.  (Generating a regexp disjunction back from this array will be
easy, a small part of the deleted code did just that.)