Add CacheWithDisposal to cache resources that require disposal (#1044)
* Add CacheWithDisposal to cache resources that require disposal

Whenever an element is fetched from the cache, a value is returned along with a deref function
that must be called to allow the element to be returned.  Once evicted, the object can only be
disposed of after all its references have been returned to the cache.

The implementation is mostly just mixing a reference count onto each entry in the cache.  But,
no existing cache seems capable of bumping a reference count *before* returning a cache entry,
which would be needed for concurrent reference counted access.  Instead, keep reference counts
by making the entire cache access section single-threaded.

And now, a rant.

Caching resources that require disposal (such as anything with a file descriptor) is difficult
in garbage-collected languages, _because_ of garbage collection.  Only memory presure triggers
GC, so it is entirely inappropriate for handling other resources.  But the lack of destructors
leaves no way to force resource cleanups.  Instead we have to return an explicit function that
returns the object to the pool.  A language such as C++ calls destructors at clear places, and
would allow returning special object references that handle lifetimes.  In fact it already has
`shared_ptr` in the standard library.

* [CR] Shard CacheWithDisposal

SingleThreadedCacheWithDisposal can only run a single creation callback at a time.  Since it
will be used for async operations with S3 this will block.  Shard the cache, to stripe calls
and therefore their blocking.

This is far from perfect (the required number of shards has to be on the order of the square
of the number of cache keys in concurrent access), but it *is* cheap.

* [CR] logs: inject a log in params and avoid logging values

* [CR] fix lint checks: explicitly ignore ignorable return value

* Use type-sensitive keys for sharding

Use format string `%+#v` for shard key: this includes struct details _and_ field names, and
should better distinguish between similar types if on the same cache.