Refactor graveler_mock_test into an exported testutil/ subdir (#1090)
* Refactor graveler_mock_test into an exported testutil/ subdir

Allows using it in packages in graveler/...

* golangci-lint fixes

* Mark and skip broken test cases

Unrelated (probably) to this change.  A bug in their use of `deep.Equal` meant that they were
not testing anything, and the underlying code (in fake or SUT) was broken: Go has simplicity,
safety, and explicitness built-in.  `deep.Equal` therefore cannot see unexported fields so it
gave the false impression that all structs matched.