Cache node_modules between runs (#1097)
* Cache node_modules between runs

See
[the manual](https://docs.github.com/en/free-pro-team@latest/actions/guides/caching-dependencies-to-speed-up-workflows)
for details of how the GitHub actions cache works.

* Cache fetched Go packages and the Go build cache between runs

* Copy cache actions to nessie

nessie is the other slow workflow.  nessie and go won't share caches because they are
different workflows.

Unfortunately this is code duplication.  So just mark it in the hope we will remember
to change both together.  Currently there is no way to share steps or steps templates
between multiple GitHub workflows.  And it is _not_ (currently) possible to create an
action that wraps another action.