Generate BatchWriterClosers as needed, not via MetaRangeManager (#1152)
There is no dependency on RangeManager.  Removing it makes it easier to write a
MetaRangeManager and *improves* the current tests -- which can now use a proper
BatchWriterCloser rather than a mock.

Remove TestWriter_SortOnClose: it tests some correctness.  But it is quite hard
to test correctly, and ends up only verifying that one `sort.Slice` line exists
in the code.  Better test this during integration.