Share a single Pebble SSTable cache between all SSTable readers (#1179)
* Share a single Pebble SSTable cache between all SSTable readers

Fixes #1177 (probably): the single cache is never finalized and should be OK.

Configure a pebble SSTable cache and use _that_.  This lets the admin control the amount of
RAM used to cache SSTable blocks.  Using a single cache makes *more* sense than giving each
reader its own cache: reading a block from an open reader costs the same no matter which it
is, so a single cache maximizes throughput.  There is no design goal to minimze latency per
reader (and this would *not* translate to minimizing latency per operation, rather it would
lower throughput on ranges that are heavily used).

This single cache has 2 references, one from the metarange readers cache and the other from
the range readers cache.  Because we never release those caches the Pebble SSTable cache is
never finalized until the program stops running.  (A panic is *still* possible then; fixing
that would require writing a finalizer.  This noncomposability of pebble SSTables indicates
just how hard that is to get right.)

There may well still be SSTable reader leaks which caused the original bug.  Here we merely
prevent those leaks from failing lakefs with a panic.

* [bugfix] committed.Iterator might not close its ranges iterator

This can _also_ cause #1177: an iterator is not closed, so its cache derefer is never called.
Not clear however how it would cause the bug-as-observed.

* [bugfix] close iterators in committedManager.{Get,Apply}

And these could _also_ cause #1177: an iterator is not closed, so its cache derefer is never
called.  Not clear however how it would cause the bug-as-observed.

* [CR] Correctly handle pebble SSTable cache references

Same end-result, just more relevant reference counting on each.  Allows adding "Close"
routines that will work conveniently.  *NOT* adding such methods as there is no useful
location from which to call them.