Correct clean existing cache files during startup (#1237)
* Correct clean existing cache files during startup

Also improve warnings on immediate rejection from cache.

Fixes #1183.

* [CR] Use tfs.eviction member where possible

* [CR] Fix workspace cleanup on startup

* [CR] Keep deleting files when ristretto cache.Set() fails

A synchronous failure is due to the ristretto worker goroutine not being set up
correctly (possibly during cache shutdown), so onReject might not be called.

But continue to log nicely, of course, to debug such weird behaviour.

* Keep evidence of failed tier_fs tests

This is a temporary directory that is left behind on test failure.  In CI it does not matter
because the container dies anyway; on a developer machine it helps.

Also improve test error detection in TestStartup.