Add OpenAPI endpoints (#1639)
* Add OpenAPI endpoints

Allow splitting "upload" support for (thick) clients:

1. Client requests a physical location and an opaque token to upload a key to a branch on a
   repo.
2. Client uploads object to physical location.
3. Client requests _linking_ the physical location to the key and completes the upload.  If
   the requested token does not match the existing token then a commit occurred, and client
   should retry, possibly with a different physical location.

* Include object metadata when linking object

This is known to the client (from uploading) so no need for the server to have to query it
from the backing store.  We can revisit and make it optional at some later date if we have
to support currently unknown client use-cases.