Wrap Python ApiClient to allow uploading most IOBase objects (#2244)
Generated Python OpenAPI code needs upload contents to be File's... but that
is merely to try to guess a MIME type from them.  Wrap ApiClient (that makes
this decision) to patch other IOBase objects with an arbitrary name (with no
extension, so safe no guessable MIME type).

This is easier than writing and using alternative Python templates that will
avoid placing filenames on uploaded objects.  Also the RFC isn't clear to me
whether a filename should go there; see #2241 for some discussion of this.