Add a forwarding ClassLoader to help load 2 copies of rocksdbjni (#2225)
* Initial implementation of AsInterfaceClassLoader

Loads from _another_ "remote" ClassLoader, places desired types into
a (presumably "local") interface type.  This lets you read any types
from the remote ClassLoader and use them -- create a new object with
`newInstance` and cast and use it as the local interface type.

Translate all types using the interfaces Map (but no more).  This is
still not enough to handle real return or parameter types -- a later
commit will add a MethodVisitor that does that.

* Configure full stacktraces in tests, conditionally dump ASM results

Add an optional PrintWriter argument to AsInterfaceClassLoader.  If
non-null, everything will be dumped there.

* Add overloaded casting methods

Works for existing tests... which do not yet actually use an overloaded
casting method.

* Verify all interface methods exist during bytecode translation

This is still a dynamic (runtime) test, but at least it happens at bytecode
translation ("loadClass") time rather than when first using a missing
method.

* Remove DEBUG prints

* Add readable constructor calls

* sbt scalafmt

* Remove UnimplementedMethodsException, it is not possible to detect it

Calls to methods of the interface that are not implemented will only be
detected when called, as AbstractMethodError.

* Change AICL.newInstance to return class based on generic argument

It generates an AnyRef that _anyway_ will need to be cast in order to be
used.  It's shorter and no more dangerous to pass a generic compile-time
argument to newInstance.

* scalafmt

* Drop leftover import

* Generate casting methods from interface

Cannot use code-to-translate to generate methods:

1. It does not include inherited methods.
2. It includes methods that we do not wish to translate.

* sbt scalafmt and unused import

* [CR] Use ASM Type.getOpcode to generate LOAD and RETURN instructions

* [CR] Remove unnecessary "equals" method; add comments

* sbt scalafmt

Ouch, it uses the worst horrible Scaladoc style.