Bugfix/2400 change error message repo with new storage type (#2469)
Changed error handling when trying to access the storage in create repository. 
In case of error with parsing the repository url or in case of miss-match of the storage type - we return a descriptive error message. In any other case we return a general error message indicating we failed to access the storage.