(Optionally) allow authentication on a configured LDAP server (#2525)
* Refactor builtin username/password authentication into auth service

* Add LDAP login

Works... but authorization then fails: current authorization relies on
mapping the JWT to an _access key_ not a _user identity_.

* Use user ID not access key ID as principal on JWT token

Now LDAP and local logins work.

* Create new LDAP users on configurable group

* Set default default LDAP user group to Viewers

* Document LDAP authentication

Also some code cleanups.

* Add friendly_name, source columns to auth_users

* Use API login to generate JWT token when testing auth

It is now too hard to generate a correct JWT token from a client -- need a
user ID.  So just log in.  (It remains easy enough to generate an incorrect
JWT token, so fake one directly that can be validated as a token for some
nonexistent user.)

* Fill in source and "friendly name" for created users

* Fix linter warnings

* Display user "friendly name" in GUI

Avoid using the full user DN to describe the user to themselves, use instead
the identifying part of the user object on LDAP.  This might be non-unique
(if the filter for finding users on LDAP changed, or if users changed on the
LDAP server), but it *is* good enough to tell the user who they are.

* [CR] Rename newLDAPAuthenticator, prevent log-flooding on login fail

Also cleanup and fix type of BaseSearchRequest.