Re-open control connection after it closes, and add timeouts (#2613)
* Re-open control connection after it closes, and add timeouts

Add much shorter LDAP timeouts to the control connection so users do not
need to wait for minutes to fail to login.

The control connection may close at any time, because TCP.  When it does
re-open it on the *next* connection attempt.  (The first attempted login
after being disconnected may fail.  I consider this a feature because it
will rarely happen, the user will try again to login, and if the attempt
just made is what caused the server to fail then we should give the user
some feedback.)

Fixes #2606.

* Name LDAP timeouts

Not on configuration because of impossibility to set default values on
optional subobjects when using viper.  Defer on this until a real user
requirement for this shows up.