Test repartitioning on all data formats in lakefsfs (#2814)
Add tests of repartitioning on all Spark data formats. Also improve compatibility with e.g. JSON, which orders columns randomly, and with CSV, which needs to write and read header lines in order to have a schema.

Relevant to #2810 testing, but does not resolve it.

* Repartition data in test

* Always test and report all formats

Even if a format fails, test the others too.  The extra runtime is marginal
compared to setup time, so the added info makes it worth it.

* Explicitly request "overwrite" save mode

* Repartition just once

* Use named columns to verify DataFrame

Columns are not ordered in JSON.

* Report exceptions per-format in Sonnets test

Allows identifying failures by format.

* Re-read CSV format files with headers as schema