Correctly prefix "after" params in lakeFS auth service (#4353)
* Correctly prefix "after" params in lakeFS auth service

Contains tests _only_ for users-by-prefix.  Might be useful to test for
everything-by-prefix, each call is subtly different.

* Fix linter errors

(Harmless)

* Fix test: amount==1 does not work

An API issue with amount==1 prevents pagination (there is also a **bug**
generating has-next in that case, but even if we fixed that the API issue
would still prevent clients from being able to page).  So don't test that.

Opened #4356 to handle that API issue.