[metadata client] Start scalacheck-ing jpebble (#4955)
* Add minimal scalacheck

Start testing blockparser using scalacheck: provide property specifications
for readInt32 and automatically generate test data.  This is not instead of
manually writing tests!  Rather, it probes code using random data following
some rules.

* Get rid of readFixedInt, it's equivalent to readInt32

Yay scalacheck, easy to gain confidence that the two are equivalent even for
Ints >= 2^31 (and therefore actually negative).

* Add spec for readMagic

And... it fails!  readMagic is broken when it reads too few bytes.

* [bug] Verify readMagic reads all the bytes it should

(Bug discovered by scalacheck!)

* scalafmt

* Downgrade scalatestplus dependencies for 2.11 support

Part of the price of supporting Spark 2.4.7 is the requirement to use Scala
2.11.  This in turn prevents using a recent testing library.