Remove metrics dynamo_{request_duration_seconds,failures_total} (#6126)
* Remove metrics dynamo_{request_duration_seconds,failures_total}

They are already covered by kv_request_duration_seconds,
kv_request_failures_total.  The first in particular is a histogram, so many
data streams.

Fixes #6125.

* Improve histogram buckets for kv_request_duration_seconds

The decade subsequence 0.1, 0.2, 0.25, 0.5, 1.0 makes little sense: its
multiples are 2, 1.25, 2, 2.  And we pay for it with reduced precision in
the lowest decade.  Instead consistently use 0.1, 0.25, 0.5, 1 in each
decade, which has multiples 2.5, 2, 2.

* Remove unused "operation" in handleClientError

* [CR] Remove another unused operation