[spark metadata client] Shade scalapb and bump its version (#7294)
* [spark metadata client] Shade scalapb and bump its version

DataBricks Spark versions now load protocol buffers... and the incompatible
version broke the lakeFS assembly.  Shade to avoid it.

Fixes #7289.

Will need to test:

* [ ] Spark standalone (part of CI)
* [ ] DBR 10.x
* [ ] DBR 12.x
* [x] DBR 13.x: _tested on 13.2 2023-01-16_.

* Set "RC" version for testing

* Actually _advance_ the version for RC

* Remove timestamp.proto

scalapb already provides it.  And it now has some weirdly different version.
This is _in any case_ a third-party file, so the less we touch it -- the
better!

* Bump prerelease version