Fix S3 gateway cross-repo copies (#7468)
* Fix S3 gateway cross-repo copies

putobject and friends were confusing src and dst repositories, which will
not do.

Fixes #7467.

* Test S3 gateway multipart copy in Esti

* [esti] Use sigV4 for S3 gateway MPU copy tests

We might only have a problem with sigV2, which is considerably less
important by now.

* Add copy-source to sigV2 signed headers and restore sigV2 signing

This should make MPU copies that are signed using sigV2 work with the S3
gateway.

* [bug] Make randomReader return EOF when it's done

Otherwise infinite loop on upload >:-(

* Try again with sigv4

* [bug] Actually request multipart-range in copy

Exercise another code path, don't triple data size, and keep the originally
intended logic.

* [bug] Use distinct repository name from TestS3CopyObject

Repositories not always deleted, so if we don't do this we get a conflict.

* [bug] Start, End on MinIO client CopySrcOptions are inclusive bounds

* [bug] Azure copyPartRange should stage block on destination

copyPartRange should stage a block on the destination.  This matters When
source and destination containers are different.

* [CR] Put all "const ...ContentLength..." together and document them

* [CR] Rename RandomReader -> NewRandomReader

Because it returns a new random reader, and also that matches the Go naming
conventions. :-)