Count KV retries not tries in Graveler (#7541)
* Count KV retries not tries in Graveler

Metric `graveler_kv_retries` currently counts "tries", i.e. at least one for
each try.  So this includes the actual operations rate: if lakeFS performs
100 ops/sec, then it will increment graveler_kv_retries by _at least_ 100 /
sec.  That defeats using this metric to detect when KV retries impact
performance.

Instead count from 0, i.e. count "_retries_".

This is changelog-worthy because people may be tracking this.  But given
that it's an internal metric, and that we rarely examine it, this is not a
breaking change.

* [CR] monitorRetries should take a "retries" parameter

* [CR] Fix silly bugs in previous commit

That's not how you subtract 1.

* [CR] Correctly name "retries" field