Add more context-based logging (#7804)
* Add context to JWT logs

Connects ID generation to its request.

* Add context to micro-batching

Store the _first_ request context with each batch.  Connects batching to its
_first_ request.  Other requests remain hard to track.

* Mark S3 AWS logs with the S3 block adapter context

We cannot log the request context with logs from AWS SDK because it does not
transfer external context.  This only marks the logs are relevant to the
adapter component.  It will become more important when we allow multiple
block adapters.