Trim  patch versions from lakeFSFS / lakeFS compatibility tests vector (#8300)
* List lakeFS versions on separate lines of lakeFSFS compatiblity test

It's getting _really_ long.  Also correct the comment about what being on
the list actually means.

Relevant to #8299.  Resolves one portion of it.

Rebased to include 1.40.0.

* Test only the latest patch version of every minor 1.x release

For any minor version, only the latest patch version needs to keep running
with with lakeFS.  After all: suppose a new lakeFSFS version does not work
with a lakeFS minor version.  If this were an issue, then we would release
some patch release for that lakeFS version.

This brings lakeFSFS support closer to that of any other lakeFS client: we
keep compatibility through the API and to fix any server-side breakages, a
new patch release.

Part of #8299, but not all of it.