Switch maven publishing portal from ossrh to new maven central (#9214)
* Bump SBT version

1.11.0 is the first version documented to support Maven Central.

* Switch Maven publisher, and secrets from OSSRH_* to MAVEN_*

Use the new Maven publisher for new Maven Central.  This commit _assumes_
that the old SBT publisher will work with new Maven Central - after all the
underlying protocol did not change AFAIK.

Fixes #9213 (apart from actually _publishing_ something).

* Also publish SBT (Spark metadata client)

* Explicitly configure Maven central endpoint

This _usually_ goes in settings.xml, but that does not work well for our CI.

* Add settings for "central" portal to publish (mvn; Java)

settings.xml is the standard place to set the central portal.

* Don't wait until published

That takes a _long_ time (>13 minutes in the one test I ran).  Running a
poll
[here](https://treeverseio.slack.com/archives/C027PG0LWAV/p1750673102138319).

Restore "prepare publish credentials" step in SBT publish

We still need it.

Re-unify SBT publish to Maven Central and AWS

That way we get to assemble it once, and assemble it the way we want to.

Make some "publish" targets phony

[bug] SBT needs to publishSigned before it can publish to Sonatype

...because Scala and SBT.

[bug] Restore GPG credentials generation for SBT signing

d'oh.

Load SBT once

Reloading is just slow and introduces _more_ errors.

* SBT: Upload to S3 before Maven Central

S3 is faster... and apparently more used as it's the assembly.

* Fix AWS auth

* Use environment variable names expected by SBT Sonatype Central publisher

* Remove "CHANGE-ME" fake release tag from SBT

It was used only for testing.

* [CR] Clean up YAML and some weirdness in .sbt/credentials